+++ src/Exception/AutowireException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SubstancePHP\Container\Exception;
+
+use Psr\Container\ContainerExceptionInterface;
+
+class AutowireException extends \RuntimeException implements ContainerExceptionInterface
+{
+}